+++ packages/layout/src/components/SiderMenu/BaseMenu.tsx
@@ -4,6 +4,7 @@ import type { MenuProps } from 'antd';
 import { Menu, Skeleton } from 'antd';
 import type { ItemType } from 'antd/es/menu/hooks/useItems';
 import classNames from 'classnames';
+import warningOnce from 'rc-util/lib/warning';
 import React, { useContext, useEffect, useMemo, useRef } from 'react';
 import type { LayoutDesignToken } from '../../context/ProLayoutContext';
 import { ProLayoutContext } from '../../context/ProLayoutContext';
@@ -149,6 +150,7 @@ class MenuUtil {
 
     const name = this.getIntlName(item);
     const children = item?.children || item?.routes;
+    warningOnce(!item?.routes, 'routes 将会废弃，为了保证兼容请使用 children 作为子节点定义方式');
     const menuType = isGroup && level === 0 ? ('group' as const) : undefined;
 
     if (Array.isArray(children) && children.length > 0) {